Use `NameGenerator` for composite name generation
The generator is already used for MR name generation.

Changes:

* since used by two controllers, `NameGenerator` moved under `internal/names` package
* `configurator_test.go` updated
