+++ internal/controller/apiextensions/claim/configurator_test.go
@@ -18,6 +18,7 @@ package claim
 
 import (
 	"context"
+	"strconv"
 	"strings"
 	"testing"
 
@@ -43,7 +44,6 @@ func TestCompositeConfigure(t *testing.T) {
 	apiVersion := "v"
 	kind := "k"
 	now := metav1.Now()
-	errBoom := errors.New("boom")
 
 	type args struct {
 		ctx context.Context
@@ -58,7 +58,6 @@ func TestCompositeConfigure(t *testing.T) {
 
 	cases := map[string]struct {
 		reason string
-		c      client.Client
 		args   args
 		want   want
 	}{
@@ -164,70 +163,8 @@ func TestCompositeConfigure(t *testing.T) {
 				err: errors.New(errBindCompositeConflict),
 			},
 		},
-		"DryRunError": {
-			reason: "We should return any error we encounter while dry-run creating a dynamically provisioned composite",
-			c: &test.MockClient{
-				MockCreate: test.NewMockCreateFn(errBoom),
-			},
-			args: args{
-				cm: &claim.Unstructured{
-					Unstructured: unstructured.Unstructured{
-						Object: map[string]any{
-							"apiVersion": apiVersion,
-							"kind":       kind,
-							"metadata": map[string]any{
-								"namespace": ns,
-								"name":      name,
-							},
-							"spec": map[string]any{
-								"coolness": 23,
-
-								// These should be preserved.
-								"compositionRef":      "ref",
-								"compositionSelector": "ref",
-
-								// These should be filtered out.
-								"resourceRef":                "ref",
-								"writeConnectionSecretToRef": "ref",
-							},
-						},
-					},
-				},
-				cp: &composite.Unstructured{},
-			},
-			want: want{
-				cp: &composite.Unstructured{
-					Unstructured: unstructured.Unstructured{
-						Object: map[string]any{
-							"metadata": map[string]any{
-								"generateName": name + "-",
-								"labels": map[string]any{
-									xcrd.LabelKeyClaimNamespace: ns,
-									xcrd.LabelKeyClaimName:      name,
-								},
-							},
-							"spec": map[string]any{
-								"coolness":            23,
-								"compositionRef":      "ref",
-								"compositionSelector": "ref",
-								"claimRef": map[string]any{
-									"apiVersion": apiVersion,
-									"kind":       kind,
-									"namespace":  ns,
-									"name":       name,
-								},
-							},
-						},
-					},
-				},
-				err: errors.Wrap(errBoom, errName),
-			},
-		},
 		"ConfiguredNewXR": {
 			reason: "A dynamically provisioned composite resource should be configured according to the claim",
-			c: &test.MockClient{
-				MockCreate: test.NewMockCreateFn(nil),
-			},
 			args: args{
 				cm: &claim.Unstructured{
 					Unstructured: unstructured.Unstructured{
@@ -260,6 +197,7 @@ func TestCompositeConfigure(t *testing.T) {
 						Object: map[string]any{
 							"metadata": map[string]any{
 								"generateName": name + "-",
+								"name":         name + "-1",
 								"labels": map[string]any{
 									xcrd.LabelKeyClaimNamespace: ns,
 									xcrd.LabelKeyClaimName:      name,
@@ -590,9 +528,6 @@ func TestCompositeConfigure(t *testing.T) {
 		},
 		"SkipK8sAnnotationPropagation": {
 			reason: "Claim's kubernetes.io annotations should not be propagated to XR",
-			c: &test.MockClient{
-				MockCreate: test.NewMockCreateFn(nil),
-			},
 			args: args{
 				cm: &claim.Unstructured{
 					Unstructured: unstructured.Unstructured{
@@ -632,6 +567,7 @@ func TestCompositeConfigure(t *testing.T) {
 						Object: map[string]any{
 							"metadata": map[string]any{
 								"generateName": name + "-",
+								"name":         name + "-1",
 								"labels": map[string]any{
 									xcrd.LabelKeyClaimNamespace: ns,
 									xcrd.LabelKeyClaimName:      name,
@@ -658,9 +594,6 @@ func TestCompositeConfigure(t *testing.T) {
 		},
 		"SkipK8sLabelPropagation": {
 			reason: "Claim's kubernetes.io annotations should not be propagated to XR",
-			c: &test.MockClient{
-				MockCreate: test.NewMockCreateFn(nil),
-			},
 			args: args{
 				cm: &claim.Unstructured{
 					Unstructured: unstructured.Unstructured{
@@ -700,6 +633,7 @@ func TestCompositeConfigure(t *testing.T) {
 						Object: map[string]any{
 							"metadata": map[string]any{
 								"generateName": name + "-",
+								"name":         name + "-1",
 								"labels": map[string]any{
 									xcrd.LabelKeyClaimNamespace: ns,
 									xcrd.LabelKeyClaimName:      name,
@@ -726,7 +660,7 @@ func TestCompositeConfigure(t *testing.T) {
 
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
-			c := NewAPIDryRunCompositeConfigurator(tc.c)
+			c := NewAPICompositeConfigurator(&mockNameGenerator{})
 			got := c.Configure(tc.args.ctx, tc.args.cm, tc.args.cp)
 			if diff := cmp.Diff(tc.want.err, got, test.EquateErrors()); diff != "" {
 				t.Errorf("Configure(...): %s\n-want error, +got error:\n%s\n", tc.reason, diff)
@@ -737,6 +671,16 @@ func TestCompositeConfigure(t *testing.T) {
 		})
 	}
 
+}
+
+type mockNameGenerator struct {
+	last int
+}
+
+func (m *mockNameGenerator) GenerateName(_ context.Context, obj resource.Object) error {
+	m.last++
+	obj.SetName(obj.GetGenerateName() + strconv.Itoa(m.last))
+	return nil
 }
 
 func TestClaimConfigure(t *testing.T) {
